Improves LeaderboardService and dev setup

+++ leaderboards/src/main/java/com/lunatech/leaderboards/dto/leaderboarduser/MatchExpectedResult.java
@@ -1,4 +1,4 @@
 package com.lunatech.leaderboards.dto.leaderboarduser;
 
-public record MatchExpectedResult(int teamAExpectedResult, int teamBExpectedResult) {
+public record MatchExpectedResult(double teamAExpectedResult, double teamBExpectedResult) {
 }
